+++ internal/storage/toml.go
@@ -3,6 +3,7 @@ package storage
 import (
 	"bytes"
 	"fmt"
+	"github.com/alecthomas/chroma/quick"
 	"github.com/pelletier/go-toml"
 )
 
@@ -32,11 +33,17 @@ func (p *TOMLParser) ToBytes(value interface{}, options ...ReadWriteOption) ([]b
 
 	enc := toml.NewEncoder(buf)
 
+	colourise := false
+
 	for _, o := range options {
 		switch o.Key {
 		case OptionIndent:
 			if indent, ok := o.Value.(string); ok {
 				enc.Indentation(indent)
+			}
+		case OptionColourise:
+			if value, ok := o.Value.(bool); ok {
+				colourise = value
 			}
 		}
 	}
@@ -67,6 +74,14 @@ func (p *TOMLParser) ToBytes(value interface{}, options ...ReadWriteOption) ([]b
 			} else {
 				return nil, err
 			}
+		}
+	}
+
+	if colourise {
+		source := buf.String()
+		buf.Reset()
+		if err := quick.Highlight(buf, source, "toml", ColouriseFormatter, ColouriseStyle); err != nil {
+			return nil, fmt.Errorf("could not colourise json output: %w", err)
 		}
 	}