+++ internal/storage/yaml.go
@@ -3,6 +3,7 @@ package storage
 import (
 	"bytes"
 	"fmt"
+	"github.com/alecthomas/chroma/quick"
 	"gopkg.in/yaml.v2"
 	"io"
 )
@@ -81,6 +82,17 @@ func (p *YAMLParser) ToBytes(value interface{}, options ...ReadWriteOption) ([]b
 	encoder := yaml.NewEncoder(buffer)
 	defer encoder.Close()
 
+	colourise := false
+
+	for _, o := range options {
+		switch o.Key {
+		case OptionColourise:
+			if value, ok := o.Value.(bool); ok {
+				colourise = value
+			}
+		}
+	}
+
 	switch v := value.(type) {
 	case SingleDocument:
 		if err := encoder.Encode(v.Document()); err != nil {
@@ -97,5 +109,14 @@ func (p *YAMLParser) ToBytes(value interface{}, options ...ReadWriteOption) ([]b
 			return nil, fmt.Errorf("could not encode default document type: %w", err)
 		}
 	}
+
+	if colourise {
+		source := buffer.String()
+		buffer.Reset()
+		if err := quick.Highlight(buffer, source, "yaml", ColouriseFormatter, ColouriseStyle); err != nil {
+			return nil, fmt.Errorf("could not colourise json output: %w", err)
+		}
+	}
+
 	return buffer.Bytes(), nil
 }